trying to make this library build fine inside mpir

+++ src/client.rs
@@ -137,7 +137,7 @@ impl<'a> PirClient<'a> {
 
             // offset into the FV plaintext
             let offset = get_fv_offset(self.client, ele_index, self.ele_size);
-            assert!(offset + self.ele_size <= result_size as u32);
+            assert!(offset + self.ele_size <= result_size);
 
             let r = slice::from_raw_parts_mut((ptr as *mut T).offset(offset as isize), 1).to_vec();
             libc::free(ptr as *mut libc::c_void);
